fix(trusty-search): shared path normalization + disk-existence guard for prune (no mass-delete on mismatch) (refs #848)
- Extract `to_corpus_relative_path(root, path)` as the single canonical
  normalisation used by BOTH `prepare_batch_payload` (chunk writer) and
  `prune_deleted_files_from_staging` (set-difference builder). Previously
  each site had its own inline `strip_prefix(...).unwrap_or(...).display()`
  expression; a future divergence (e.g. an `unwrap_or` returning an absolute
  path on one side but not the other) would make the set-difference classify
  live files as "deleted" and wipe the staging corpus.

- Add disk-existence guard (fix 1, data-safety critical): before any
  per-file prune, reconstruct the absolute path (`canonical_root.join(relative)`)
  and verify it does NOT exist on disk via `Path::exists()`. If the file is
  still present, it is a false positive from a path-normalisation mismatch —
  the prune is skipped and a `warn!` is emitted. Only files confirmed absent
  from disk are ever removed. This makes mass-deletion from a normalisation
  bug structurally impossible.

- Batch the hash-entry delete (fix 2): collect all confirmed-pruned paths
  and call `delete_file_hash_entries(&all_pruned)` once after the loop
  (one redb write-txn instead of one per file).

- Aggregate swallowed errors (fix 3): track `failed_count` across all per-file
  operations and emit a single aggregated `warn!` at the end when > 0, so
  operators can detect persistent partial-prune failures.

- Add `// TODO` note on `list_indexed_files` full-table-scan perf (FILES_TABLE
  secondary index deferred).

- Split tests to `prune_tests.rs` (loaded via `#[path]`) to keep both files
  under the 500-line cap.

- New tests: `to_corpus_relative_path_agrees_with_batch_loop` (normalisation
  round-trip), `disk_existence_guard_skips_live_file` (guard predicate).
  All 3 existing prune tests continue to pass.
