+++ crates/trusty-search/src/service/reindex/mod.rs
@@ -19,7 +19,7 @@ pub mod quarantine;
 mod staging;
 mod validate;
 
-use prune::prune_deleted_files_from_staging;
+use prune::{prune_deleted_files_from_staging, to_corpus_relative_path};
 
 pub use quarantine::ReindexQuarantine;
 
@@ -1070,11 +1070,13 @@ async fn prepare_batch_payload(ctx: &BatchCtx, batch: &[PathBuf]) -> BatchPayloa
     let mut to_index_paths: Vec<PathBuf> = Vec::with_capacity(batch.len());
     let mut new_hashes: Vec<(PathBuf, String)> = Vec::with_capacity(batch.len());
     for (path, content_res) in read_results {
-        let rel = path
-            .strip_prefix(&ctx.root)
-            .unwrap_or(&path)
-            .display()
-            .to_string();
+        // Use the shared canonical normalisation from `prune` so every
+        // relative-path string stored in the corpus is produced by the
+        // same function that the prune pass uses when building its walked-set.
+        // This ensures the set-difference in `prune_deleted_files_from_staging`
+        // can never generate a false "deleted" entry due to a per-site
+        // divergence in how the relative path is formatted.  (Issue #854.)
+        let rel = to_corpus_relative_path(&ctx.root, &path);
         let content = match content_res {
             Ok(c) => c,
             Err(e) => {
@@ -1112,12 +1114,9 @@ async fn prepare_batch_payload(ctx: &BatchCtx, batch: &[PathBuf]) -> BatchPayloa
         // returns canonicalised paths under the root; the `unwrap_or` is a
         // defensive fallback that preserves the previous behaviour for any edge-
         // case path that somehow escapes the root (e.g. a symlink target outside
-        // the tree).
-        let path_str = path
-            .strip_prefix(&ctx.root)
-            .unwrap_or(&path)
-            .display()
-            .to_string();
+        // the tree).  `rel` above was already computed via `to_corpus_relative_path`,
+        // which uses the same strip_prefix + unwrap_or + display idiom — reuse it.
+        let path_str = rel.clone();
         to_index.push((path_str, content));
         to_index_paths.push(path.clone());
         new_hashes.push((path, h));
@@ -1154,7 +1153,7 @@ async fn emit_batch_error(ctx: &BatchCtx, to_index_paths: &[PathBuf], err: anyho
     use std::sync::atomic::Ordering;
     let files_in_batch: Vec<String> = to_index_paths
         .iter()
-        .map(|p| p.strip_prefix(&ctx.root).unwrap_or(p).display().to_string())
+        .map(|p| to_corpus_relative_path(&ctx.root, p))
         .collect();
     ctx.progress
         .errors
